backend/lambdas/requirements.txt
psycopg2-binary==2.9.9
stripe
boto3==1.35.91
anthropic==0.40.0
google-genai==1.40.0
requests==2.31.0